Continuation of Stripe & Merch updates
Continued the effort in creating a Stripe component to successfully
run on check out of a Merch item.  At this point, I'm able to use a
button to create the modal but the button to generate the modal won't
appear on the page that I want it to and generates an error every
time I try.  Outside of that, I simplified the CSS a bit and was able
to resolve my Show & Index issues to run successfully.